Making it build after API changes
