+++ Source/Kernel/Grains.Specs/Events/Constraints/for_ConstraintValidationContext/when_updating.cs
@@ -5,12 +5,12 @@ using Cratis.Chronicle.Concepts.Events;
 
 namespace Cratis.Chronicle.Grains.Events.Constraints.for_ConstraintValidationContext;
 
-public class when_updating : given.a_constraint_validation_context_with_two_validators
+public class when_updating : given.a_constraint_validation_context_with_two_validators_that_are_index_updaters
 {
     static readonly EventSequenceNumber _sequenceNumber = 42;
 
     async Task Because() => await _context.Update(_sequenceNumber);
 
-    [Fact] void should_call_the_first_validator() => _firstValidator.Received(1).Update(_context, _sequenceNumber);
-    [Fact] void should_call_the_second_validator() => _secondValidator.Received(1).Update(_context, _sequenceNumber);
+    [Fact] void should_call_the_first_validator() => _firstValidatorIndexUpdater.Received(1).Update(_sequenceNumber);
+    [Fact] void should_call_the_second_validator() => _secondValidatorIndexUpdater.Received(1).Update(_sequenceNumber);
 }
